+++ OrchEngine/src/main/java/cc/abro/orchengine/gameobject/components/render/SpriteRender.java
@@ -16,7 +16,7 @@ public class SpriteRender extends Rendering {
 
     @Override
     public void draw() {
-        Vector2<Integer> relativePosition = getGameObject().getRelativePosition();
+        Vector2<Double> relativePosition = getGameObject().getRelativePosition();
         double xView = relativePosition.x;
         double yView = relativePosition.y;
         double directionDraw = getGameObject().getDirection();
